+++ pylib/terms.py
@@ -13,6 +13,7 @@ ONYA = I('http://purl.org/onya/vocab/')
 ONYA_TYPE_REL = ONYA('type')
 ONYA_SOURCE_REL = ONYA('source')  # @source provenance sub-property
 ONYA_DOCUMENT = ONYA('Document')  # Type for document nodes created from @document directive
+ONYA_ASSERTION = ONYA('Assertion')  # Implicit type carried by every assertion (edge/property)
 
 # RDF